App/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.6
Jinja2==3.1.2
Flask-WTF==1.1.1
WTForms==3.0.1
itsdangerous==2.1.2
reportlab==4.0.7
